Moved ServerChannel to the server object and fixed the test which used it.
